+++ apps/manager/src/api/mutations/useCheckDuplicateNL.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@hcc/api-base';
 
 import type { CheckDuplicateNLResponse, ParsedPlayer } from '~/api/types/nl';
 
-import { fetcher } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
 
 const postCheckDuplicateNL = (payload: { players: ParsedPlayer[] }) => {
   return fetcher.post<CheckDuplicateNLResponse>(`nl/check-duplicates`, { json: payload });
