+++ apps/manager/src/api/mutations/useCreatePlayers.ts
@@ -2,7 +2,8 @@ import { useMutation, useQueryClient } from '@hcc/api-base';
 
 import type { PlayerType } from '~/api';
 
-import { fetcher, queryKeys } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
+import { queryKeys } from '~/api/queryKey';
 
 export type PlayerFormType = Pick<PlayerType, 'name' | 'studentNumber'>;
