+++ apps/manager/src/api/mutations/useCreateTimelinePK.ts
@@ -2,7 +2,8 @@ import { useMutation, useQueryClient } from '@hcc/api-base';
 
 import type { PkType } from '~/api';
 
-import { fetcher, queryKeys } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
+import { queryKeys } from '~/api/queryKey';
 
 export const postTimelinePK = ({ gameId, ...request }: PkType) => {
   return fetcher.post<void>(`games/${gameId}/timelines/pk`, {
