+++ apps/manager/src/api/mutations/useCreateTimelineScore.ts
@@ -2,7 +2,8 @@ import { useMutation, useQueryClient } from '@hcc/api-base';
 
 import type { ScoreType } from '~/api';
 
-import { fetcher, queryKeys } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
+import { queryKeys } from '~/api/queryKey';
 
 export const postTimelineScore = ({ gameId, ...request }: ScoreType) => {
   return fetcher.post<void>(`games/${gameId}/timelines/score`, {
