+++ apps/manager/src/api/mutations/useCreateTimelineStatus.ts
@@ -2,7 +2,8 @@ import { useMutation, useQueryClient } from '@hcc/api-base';
 
 import type { ProgressStateType } from '~/api';
 
-import { fetcher, queryKeys } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
+import { queryKeys } from '~/api/queryKey';
 
 export const postTimelineProgress = ({ gameId, ...request }: ProgressStateType) => {
   return fetcher.post<void>(`games/${gameId}/timelines/progress`, {
