+++ apps/manager/src/api/mutations/useCreateTimelineWarning.ts
@@ -2,7 +2,8 @@ import { useMutation, useQueryClient } from '@hcc/api-base';
 
 import type { WarningType } from '~/api';
 
-import { fetcher, queryKeys } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
+import { queryKeys } from '~/api/queryKey';
 
 export const postTimelineWarning = ({ gameId, ...request }: WarningType) => {
   return fetcher.post<void>(`games/${gameId}/timelines/warning-card`, {
