+++ apps/manager/src/api/mutations/useParseNL.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@hcc/api-base';
 
 import type { ParseNLPayload, ParseNLResponse } from '~/api/types/nl';
 
-import { fetcher } from '../queryKey';
+import { fetcher } from '~/api/fetcher';
 
 const postParseNL = (payload: ParseNLPayload) => {
   return fetcher.post<ParseNLResponse>(`nl/parse`, { json: payload });
