+++ apps/manager/src/api/mutations/useRegisterNL.ts
@@ -2,7 +2,8 @@ import { useMutation, useQueryClient } from '@hcc/api-base';
 
 import type { RegisterNLResponse, RegisterNLPayload } from '~/api/types/nl';
 
-import { fetcher, queryKeys } from '~/api/queryKey';
+import { fetcher } from '~/api/fetcher';
+import { queryKeys } from '~/api/queryKey';
 
 const postRegisterNL = (payload: RegisterNLPayload) => {
   return fetcher.post<RegisterNLResponse>(`nl/register-team`, { json: payload });
